perf: Dramatically lower the time spent in the reading routine.
The header of the command sent to the PIO code is reduced from 2 FIFO write to a single one.
The data is 2x16 bits, allowing to cover all the needs.

What used to take close to 1 millisecond now takes between 4 and 6 microseconds.
This is due to the fact that there is no more stall on the FIFO.